replace simplemde with textarea and markdown key
simplemde was not usable with smartphones (smart keyboards) and
development seems to have ceased. the content of posts are mostly plain
text and maybe a couple of links, a plain textarea + a key for common
markdown syntax could suffice? can't find any good replacements for
simplemde that does not require a full node.js/javascript environment.